Modularized code more
